+++ changelog.md
@@ -5,10 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## unreleased
 
-### Changed
-
 ### Fixed
 - Fix the trigger metadata icons failing to render by removing the license header placed before the XML declaration
+- [Apply the standard observability support for the FTP Library](https://github.com/ballerina-platform/module-ballerina-ftp/pull/1588)
 
 ## [2.20.2] - 2026-09-08